refactor(canvas): deepen chat message module
