+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessage/index.tsx
@@ -1,29 +1,16 @@
-import { useCallback, useEffect, useMemo, useRef, useState, type KeyboardEvent, type SyntheticEvent } from 'react';
 import './index.css';
 import type { AgentChatMessage, CanvasNode } from '../../../types';
 import { toFileUrl } from '../../../utils/fileUrl';
 import { BotAvatarIcon, PencilIcon, RefreshIcon } from '../../icons';
 import type { ToolCallStatus } from '../../../types';
-import { useRoleColors, useRoleNameColors } from '../../../agent-chat/mentions/roleMentionItems';
-import { renderMdWithMentions } from '../utils/mentions';
 import { roleColorSoft } from '../../../utils/roleColors';
-import { isImeComposing } from '../../../utils/ime';
-import { renderMermaidIn } from '../../../utils/mermaid';
-import { formatAbsoluteTime, formatRelativeTime } from '../utils/time';
-import { ChatToolCalls } from './ChatToolCalls';
 import { ChatActivityStatus } from '../ChatMessages/ChatActivityStatus';
-import { ChatImageLightbox, type LightboxImage } from '../ChatImageLightbox';
+import { ChatImageLightbox } from '../ChatImageLightbox';
 import { PluginChatCardForMessage } from '../../../../../plugins/renderer';
-import {
-  ChatArtifactCard,
-  ChatInlineVisual,
-  parseVisualToolResult,
-} from '../../artifacts';
-import { CopyGeneratedImageButton, parseGeneratedImage } from './GeneratedImageActions';
-import { useI18n } from '../../../i18n';
 import { ChatTurnOutcome } from './ChatTurnMeta';
 import { CopyMessageButton } from './ChatMessageActions';
-import { McpAppFramesLazy } from './McpAppFramesLazy';
+import { useChatMessageController } from './useChatMessageController';
+import { ChatMessageToolResults } from './ChatMessageToolResults';
 
 interface ChatMessageProps {
   message: AgentChatMessage;
@@ -75,147 +62,45 @@ export const ChatMessage = ({
   turnStartedAt,
   onSessionJump,
 }: ChatMessageProps) => {
-  const { t } = useI18n();
-  // Live role accents so `@角色` chips in the transcript match each role's
-  // color (falls back to the violet tokens for deleted/unknown roles).
-  const roleColors = useRoleColors();
-  // Assistant-only: chip the plain `@Name` a role writes when handing off.
-  const roleNames = useRoleNameColors();
-  const assistantHtml = useMemo(
-    () => (message.role === 'assistant'
-      ? renderMdWithMentions(message.content, nodes, { streaming: isStreaming, rootFolder, roleColors, roleNames })
-      : ''),
-    [message.role, message.content, nodes, isStreaming, rootFolder, roleColors, roleNames],
-  );
-  const userHtml = useMemo(
-    () => (message.role === 'user'
-      ? renderMdWithMentions(message.content, nodes, { rootFolder, roleColors })
-      : ''),
-    [message.role, message.content, nodes, rootFolder, roleColors],
-  );
-  // Copy is offered for any settled message (user or assistant) with a body.
-  const showCopyToolbar = !isStreaming && !!message.content;
-  const relativeTime = formatRelativeTime(message.timestamp);
-  const absoluteTime = formatAbsoluteTime(message.timestamp);
-  const speakerLabel = message.speakerRoleName
-    ?? (message.role === 'assistant' ? t('chat.assistantSpeaker') : t('chat.userSpeaker'));
-
-  const [isEditing, setIsEditing] = useState(false);
-  const [editValue, setEditValue] = useState('');
-  const [liveToolDetailsOpen, setLiveToolDetailsOpen] = useState(false);
-
-  const canEdit = message.role === 'user'
-    && !!onEditUserMessage
-    && !loading
-    && !isStreaming;
-  const canRegenerate = message.role === 'assistant'
-    && !!onRegenerate
-    && !loading
-    && !isStreaming
-    && !message.turnStatus;
-  const canRecoverTurn = !!onRegenerate && !loading && !isStreaming;
-
-  const handleStartEdit = useCallback(() => {
-    setEditValue(message.content);
-    setIsEditing(true);
-  }, [message.content]);
-
-  const handleCancelEdit = useCallback(() => {
-    setIsEditing(false);
-  }, []);
-
-  const handleSaveEdit = useCallback(async () => {
-    if (!onEditUserMessage) return;
-    const trimmed = editValue.trim();
-    if (!trimmed) return;
-    // Close the editor only when the rewind+resend actually went through —
-    // otherwise (e.g. another turn is still streaming) the user's edited
-    // text would be silently discarded.
-    const ok = await onEditUserMessage(index, trimmed);
-    if (ok !== false) setIsEditing(false);
-  }, [editValue, index, onEditUserMessage]);
-
-  const handleEditKeyDown = useCallback((event: KeyboardEvent<HTMLTextAreaElement>) => {
-    // Escape/Enter during IME composition dismiss/confirm the candidate
-    // text — cancelling the edit or saving there would eat CJK input.
-    if (isImeComposing(event)) return;
-    if (event.key === 'Escape') {
-      event.preventDefault();
-      handleCancelEdit();
-      return;
-    }
-    if (event.key === 'Enter' && (event.metaKey || event.ctrlKey)) {
-      event.preventDefault();
-      void handleSaveEdit();
-    }
-  }, [handleCancelEdit, handleSaveEdit]);
-
-  const handleRegenerate = useCallback(() => {
-    if (!onRegenerate) return;
-    void onRegenerate(index);
-  }, [index, onRegenerate]);
-
-  const handleImageError = useCallback((event: SyntheticEvent<HTMLImageElement>) => {
-    const card = event.currentTarget.closest('.chat-message-image-card');
-    card?.classList.add('chat-message-image-card--broken');
-  }, []);
-
-  // Click-to-zoom: any image in this message (user attachments first, then
-  // generated images) opens a shared fullscreen viewer at its position.
-  const [lightboxIndex, setLightboxIndex] = useState<number | null>(null);
-
-  const generatedImages = useMemo(() => {
-    if (message.role !== 'assistant' || !tools) return [];
-    const out: Array<{ key: string; src: string; outputPath: string; title?: string }> = [];
-    for (const tool of tools) {
-      const image = parseGeneratedImage(tool.result);
-      if (!image?.outputPath) continue;
-      out.push({
-        key: `generated-${tool.id}`,
-        src: toFileUrl(image.outputPath),
-        outputPath: image.outputPath,
-        title: image.title,
-      });
-    }
-    return out;
-  }, [message.role, tools]);
-
-  const attachmentCount = message.attachments?.length ?? 0;
-
-  const lightboxImages = useMemo<LightboxImage[]>(() => [
-    ...(message.attachments ?? []).map(attachment => ({
-      src: toFileUrl(attachment.path),
-      filePath: attachment.path,
-      caption: attachment.fileName,
-    })),
-    ...generatedImages.map(image => ({
-      src: image.src,
-      filePath: image.outputPath,
-      caption: image.title,
-    })),
-  ], [message.attachments, generatedImages]);
-
-  const handleImageKeyOpen = useCallback(
-    (event: KeyboardEvent<HTMLImageElement>, openIndex: number) => {
-      if (event.key === 'Enter' || event.key === ' ') {
-        event.preventDefault();
-        setLightboxIndex(openIndex);
-      }
-    },
-    [],
-  );
-
-  // After every (re-)render of a message body, render any pending mermaid
-  // placeholders. Applies to both assistant output and pasted user content
-  // (a user can paste a ```mermaid block too). We skip while streaming
-  // because partial diagrams always fail to parse — they get picked up
-  // once the stream completes.
-  const bodyRef = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    if (isStreaming) return;
-    renderMermaidIn(bodyRef.current);
-  }, [assistantHtml, userHtml, isStreaming]);
-
+  const {
+    absoluteTime,
+    assistantHtml,
+    attachmentCount,
+    bodyRef,
+    canEdit,
+    canRecoverTurn,
+    canRegenerate,
+    editValue,
+    generatedImages,
+    handleCancelEdit,
+    handleEditKeyDown,
+    handleImageError,
+    handleImageKeyOpen,
+    handleRegenerate,
+    handleSaveEdit,
+    handleStartEdit,
+    isEditing,
+    lightboxImages,
+    lightboxIndex,
+    liveToolDetailsOpen,
+    relativeTime,
+    setEditValue,
+    setLightboxIndex,
+    setLiveToolDetailsOpen,
+    showCopyToolbar,
+    speakerLabel,
+    userHtml,
+  } = useChatMessageController({
+    message,
+    index,
+    isStreaming,
+    loading,
+    tools,
+    nodes,
+    rootFolder,
+    onEditUserMessage,
+    onRegenerate,
+  });
   return (
     <div
       className={`chat-message chat-message-${message.role}`}
@@ -276,116 +161,24 @@ export const ChatMessage = ({
         />
       )}
       {message.role === 'assistant' && tools && tools.length > 0 && (
-        <>
-          <ChatToolCalls
-            tools={tools}
-            collapsed={collapsed}
-            expandedTools={expandedTools}
-            showSectionHeader={!loading}
-            isStreaming={isStreaming}
-            liveDetailsOpen={liveToolDetailsOpen}
-            onToggleSection={onToggleSection}
-            onToggleToolExpand={onToggleToolExpand}
-            onSessionJump={onSessionJump}
-          />
-          <McpAppFramesLazy
-            tools={tools}
-            instanceScope={`${workspaceId}:${message.timestamp}:${index}`}
-          />
-          {generatedImages.length > 0 && (
-            <div className="chat-generated-images">
-              {generatedImages.map((image, generatedIndex) => {
-                const openIndex = attachmentCount + generatedIndex;
-                return (
-                  <figure key={image.key} className="chat-generated-image-card">
-                    <img
-                      src={image.src}
-                      alt={image.title ?? 'Generated image'}
-                      className="chat-image-clickable"
-                      role="button"
-                      tabIndex={0}
-                      onClick={() => setLightboxIndex(openIndex)}
-                      onKeyDown={(event) => handleImageKeyOpen(event, openIndex)}
-                    />
-                    <figcaption>
-                      <span>{image.title ?? 'Generated image'}</span>
-                      <span className="chat-generated-image-card__actions">
-                        <CopyGeneratedImageButton imagePath={image.outputPath} />
-                        <button
-                          type="button"
-                          className="chat-generated-image-card__action chat-generated-image-card__action--primary"
-                          onClick={() => void onAddImageToCanvas?.(image.outputPath, image.title)}
-                        >
-                          Add to canvas
-                        </button>
-                      </span>
-                    </figcaption>
-                  </figure>
-                );
-              })}
-            </div>
-          )}
-          {tools.map(tool => {
-            // visual_render in flight: drive an inline streaming preview.
-            // Prefer `streamedContent` (the tool's own side-channel chunks)
-            // when present — works on any LLM/provider — and fall back to
-            // partial JSON extraction if the upstream model genuinely
-            // streams tool args.
-            if (tool.name === 'visual_render' && !tool.result) {
-              return (
-                <ChatInlineVisual
-                  key={`visual-${tool.id}`}
-                  workspaceId={workspaceId}
-                  streamedContent={tool.streamedContent}
-                  partialInput={tool.partialInput}
-                  streaming
-                />
-              );
-            }
-            // visual_render finished but the side-channel stream may still
-            // be flushing the final frames. Until streamedDone, keep using
-            // the streaming view so the swap to the script-enabled iframe
-            // happens at the END of the animation, not on tool-result.
-            if (tool.name === 'visual_render' && tool.result && !tool.streamedDone && tool.streamedContent) {
-              return (
-                <ChatInlineVisual
-                  key={`visual-${tool.id}`}
-                  workspaceId={workspaceId}
-                  streamedContent={tool.streamedContent}
-                  streaming
-                />
-              );
-            }
-            // artifact_create / _update in flight → no inline preview at
-            // all; the tool-call chip above signals progress and the
-            // artifact card lands once the tool returns. Drawer is the
-            // right place for a live artifact preview, not the chat.
-            if (
-              (tool.name === 'artifact_create' || tool.name === 'artifact_update')
-              && !tool.result
-            ) {
-              return null;
-            }
-            const visual = parseVisualToolResult(tool.name, tool.result);
-            if (!visual) return null;
-            if (visual.kind === 'visual_render') {
-              return (
-                <ChatInlineVisual
-                  key={`visual-${tool.id}`}
-                  workspaceId={workspaceId}
-                  payload={visual.payload}
-                />
-              );
-            }
-            return (
-              <ChatArtifactCard
-                key={`artifact-${tool.id}`}
-                workspaceId={workspaceId}
-                payload={visual.payload}
-              />
-            );
-          })}
-        </>
+        <ChatMessageToolResults
+          tools={tools}
+          collapsed={collapsed}
+          expandedTools={expandedTools}
+          loading={loading}
+          isStreaming={isStreaming}
+          liveToolDetailsOpen={liveToolDetailsOpen}
+          onToggleSection={onToggleSection}
+          onToggleToolExpand={onToggleToolExpand}
+          onSessionJump={onSessionJump}
+          workspaceId={workspaceId}
+          messageTimestamp={message.timestamp}
+          messageIndex={index}
+          generatedImages={generatedImages}
+          attachmentCount={attachmentCount}
+          setLightboxIndex={setLightboxIndex}
+          onAddImageToCanvas={onAddImageToCanvas}
+        />
       )}
       {message.role === 'assistant' ? (
         isStreaming ? (
